requirements.txt
# Core graph ML
torch==2.2.0
torch-geometric==2.5.0
torch-scatter==2.1.2
torch-sparse==0.6.18
node2vec==0.4.6

# Graph database
neo4j==5.18.0

# Data & ML
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
scipy==1.12.0

# Visualization
umap-learn==0.5.6
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.20.0

# MLflow experiment tracking
mlflow==2.11.1

# API
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.3

# Utilities
python-dotenv==1.0.1
kaggle==1.6.6
tqdm==4.66.2
networkx==3.2.1

# Testing
pytest==8.1.1
pytest-cov==5.0.0
httpx==0.27.0
